Add gcc attribute cleanup support
so I've use macro to generate the call, because it was easier that
generate the call normally.

The major difficulty was to handle cleanup when a goto happen
to do so, I've had a "ScopeTracker".
I can't use local_scope because that would not work with code like below
as local_scope would be at the same level:

{
    char * __attribute__ ((cleanup(clean_function))) str = "hej";
    goto next;
}
{
    next:
}